requirements.txt
alembic==1.16.1
altgraph==0.17.4
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
chardet==5.2.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.14
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Admin==1.6.1
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
login==0.0.6
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
pyodbc==5.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pywin32==310
pywin32-ctypes==0.2.3
reportlab==4.4.1
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
waitress==3.0.2
Werkzeug==3.1.3
wkhtmltopdf==0.2
WTForms==3.2.1
xlrd==2.0.1
